+++ example/TestLinkExample.py
@@ -616,7 +616,7 @@ print "create new version of TC B"
 response = myTestLink.copyTCnewVersion(newTestCaseID_B, 
                 summary='new version of TC B', importance='1')
 print 'copyTCnewVersion', response
-print "copy TC B as TC BA into Test suite B"
+print "copy TC B as TC BA into Test suite A"
 response = myTestLink.copyTCnewTestCase(newTestCaseID_B, 
                 testsuiteid=newTestSuiteID_A, testcasename='%sA' % NEWTESTCASE_B)
 print 'copyTCnewTestCase', response